requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
